backend/requirements.txt
flask==2.0.1
flask-jwt-extended==4.3.1
flask-cors==3.0.10
python-dotenv==1.0.0
Pillow==10.0.0
tensorflow==2.15.0
numpy==1.24.3
boto3==1.18.44
werkzeug==2.0.1
pytest==6.2.5
black==21.7b0
flake8==3.9.2
python-magic==0.4.24
requests==2.31.0
pymongo==4.6.1
python-jose==3.3.0 